crypto-exchange/producer/requirements.txt
websocket-client==1.8.0
kafka-python==2.0.2
prometheus-client==0.20.0
